Add basic project setup and a demo bot
The bot starter just kicks off certain bot based on passed args.

Responder bot is supposed to plug into different discord servers and
channels and generate auto-response for things. Only thing there right
now is the rep handler that gives !cookie to anyone who gives !rep to
the running user.

Including .idea files in a repo for the first time with hopes that it
will make a fresh Intellij setup easier in future.