Add spawn listener
Spawn listener is just a JDA event handler which acts as the producer
of data for the spawn exchange. It has a set of processors that it
uses to parse spawn information in its message event handler. The
running user of the JDA is expected to have access to all the servers
for which we have the processors. It submits new spawns to the spawn
exchange.