Implement crawler for radar maps
Two maps for CL and PN state in USA are using the same server/UI.
Implementing for both of them under the name 'radar' map. Although
current impl is not working for PN, since it needs some special
setup for server certificate to work.

Also fetch queries for these maps are tailored to only fetch things
that might be of interest. Without putting this filtering we'd end
up fetching way too much data.

Only plugging in the CL map in the crawler for now.