Remove candy details from spawn stats message
With the candy details the embed message we sent over discord grows
beyond the 6000 character limit. Removing the candy details for now,
although this may not be the most ideal way to handle the limitation
since message can exceed limit due to other reasons. This is good enough
for now though.