Plug-in new map Ash Club covering some part of FL
In the region enum set, added a new region FL. This corresponds to Ash
Club's map as far as Radar crawler is concerned. Actually Ash Club map
only covers a few little bits of FL, but given how the maps today are
named after regions, added this as FL region.