Enable low-confidence despawn time parsing
Needed for one of the new source we need to add. Current implementation
looks for despawn time in one line at a time and also looks for specific
emojis, like check marks, times etc., which indicates that the time info
is indeed a despawn time.

New source has the indicator in a different line so the per line check
doesn't work for it. Just allowing "low confidence" parsing for now,
which basically skips that check. Better implementation would be being
able to process info from multiple lines at a time instead of dumbing
down any check.